# Async sequence writer