# Issue

This PR introduces `sequence_writer<T>` that offloads filesystem flush onto a separate thread.

During the profiling I've noticed pauses, caused by `sequence<T>::push_back()` and this PR is an attempt to eliminate such pauses in parsing by offloading writing onto a separate thread.

<img width="1508" alt="image" src="https://github.com/user-attachments/assets/dd80c1f4-8a1b-4ecc-a300-2ea7d46b2944" />

On AWS EC2 m7a.4xlarge with 1000 MB/s throughput and 12k IOPS EBS storage

| timing | master | this PR |
| --- | --- | --- |
| pbfgraphparser.cc::clarify_and_fix | 829s | 834s |
| pbfgraphparser.cc::ParseWays | 4252s | 4157s |
| osmdata.cc::write_to_temp_files | 2s | 2s |
| pbfgraphparser.cc::ParseRelations | 291s | 292s |
| osmdata.cc::write_to_temp_files | 4s | 5s |
| pbfgraphparser.cc::ParseNodes | 7881s | 7731s |
| osmdata.cc::write_to_temp_files | 5s | 5s |
| util.cc::build_tile_set | 12443s | 12199s |

```sh
#!/bin/bash

# git clone --recurse-submodules https://github.com/valhalla/valhalla.git
git clone --recurse-submodules https://github.com/kinkard/valhalla.git --branch async-sequence-writer

cd valhalla

./scripts/install-linux-deps.sh 

mkdir build && cd build
cmake .. -DCMAKE_BUILD_TYPE=Release -DENABLE_SINGLE_FILES_WERROR=OFF
cmake --build . -j 16

sudo snap install aws-cli --classic
aws s3 --no-sign-request --no-progress cp s3://osm-planet-eu-central-1/planet/pbf/2025/planet-250421.osm.pbf ./

mkdir valhalla_data

./valhalla_build_config \
--mjolnir-tile-dir=./valhalla_data \
--mjolnir-tile-extract=./valhalla_data/tiles.tar \
--mjolnir-admin=./valhalla_data/admins.sqlite \
--mjolnir-timezone=./valhalla_data/timezones.sqlite \
> ./valhalla_data/config.json

./valhalla_build_timezones > ./valhalla_data/timezones.sqlite

./valhalla_build_admins -c ./valhalla_data/config.json *.osm.pbf

./valhalla_build_tiles -c ./valhalla_data/config.json --end parsenodes *.osm.pbf
```

## Tasklist

 - [x] Add tests
 - [ ] Add #fixes with the issue number that this PR addresses
 - [ ] Update the docs with any new request parameters or changes to behavior described
 - [x] Update the [changelog](CHANGELOG.md)
 - [ ] If you made changes to the lua files, update the [taginfo](taginfo.json) too.

## Requirements / Relations

 Link any requirements here. Other pull requests this PR is based on?