@kevinkreiser 2025-05-01
can you update the description to explain why we want such a structure? i think i get the idea, that flush to file system takes time and if we do it on a background thread we dont have to wait to keep writing but i never would have guessed that that actually mattered in practice (unless you are on ancient spinning rust). what is the real world impact of this?