@kinkard 2025-05-02
> can you update the description to explain why we want such a structure? i think i get the idea, that flush to file system takes time and if we do it on a background thread we dont have to wait to keep writing but i never would have guessed that that actually mattered in practice (unless you are on ancient spinning rust). what is the real world impact of this?

Sure! I've updated the description of the PR. TLDR - I've noticed pauses caused by `flush()` and my main idea was that due to EBS latency it impacts parsing performance noticeably.
Sadly, the real difference on the AWS instance is barely noticeable and probably the added complexity doesn't worth it.

Another idea I had is to eventually split `sequence<T>` into two pieces - one that pushed data via `fstream` and another that provides an access via `mmap`. Sadly, there are several places where both types of access is required and it is hard to complete such separation.