@kevinkreiser 2025-05-02
> Sadly, the real difference on the AWS instance is barely noticeable (around 2% overall) and probably the added complexity doesn't worth it.

i agree, i dont think the complication is worth it.

instead it seems to me there are other places we can spend our energy to speed things up.

figure out how to have multiple threads writing to different sequences in the parsing step and then merge sort them back into a single sequence once they are all finished at each parser step. but only spend time on that if that single threaded writer is the actual bottleneck in the parsing.

what i imagine is the bottleneck in the parsing is the actual lua and c++ tag transformations themselves. at the moment that is also single threaded. i have to re-educate myself on everything in there but i do think most of that stuff (the tag transformations at least) can be done async. its too bad osmium doesnt have a way to push the transformations down into its threaded reading and instead of returning us its objects it could return is our own objects after the tag transforms are run.

aside from that the other big one is the density calculation in enhancer. imho we could add a per-tile cache to that so that we can reuse partial results from the computation. it basically just says hey sum all the lengths of the edges in this area. we could simply keep an unordered map of node_graphid -> length (of edges connected) which we fill once with a single iteration over the given tiles data. from there we're just saying for a given graph node, which other graph nodes are within N km. sum their edge lengths from the map. to speed up the spatial part where you want all the nodes within a radius from a node we can spatially sort the nodes (i forget if they are already at this point) and then binary search to find the first one that is in range on the left side of the range and quit after we hit the right side of the range (nodes are sorted like a classical scanline algorithm).