@kinkard 2025-05-02
> i agree, i dont think the complication is worth it.

👍 

> what i imagine is the bottleneck in the parsing is the actual lua and c++ tag transformations themselves. at the moment that is also single threaded. i have to re-educate myself on everything in there but i do think most of that stuff (the tag transformations at least) can be done async. its too bad osmium doesnt have a way to push the transformations down into its threaded reading and instead of returning us its objects it could return is our own objects after the tag transforms are run.

I did some improvements in that direction in the https://github.com/valhalla/valhalla/pull/5143 but two synchronization points (one where buffers are read from libosmium and second one when chunks of transormed ways fed to the C++ part of the transformation) eat quite a lot. And real measurenments on AWS EC2 instance show that `PBFGraphParser::ParseWays()` hardly able to load even 5 CPUs.

> aside from that the other big one is the density calculation in enhancer. imho we could add a per-tile cache to that so that we can reuse partial results from the computation. it basically just says hey sum all the lengths of the edges in this area. we could simply keep an unordered map of node_graphid -> length (of edges connected) which we fill once with a single iteration over the given tiles data. from there we're just saying for a given graph node, which other graph nodes are within N km. sum their edge lengths from the map. to speed up the spatial part where you want all the nodes within a radius from a node we can spatially sort the nodes (i forget if they are already at this point) and then binary search to find the first one that is in range on the left side of the range and quit after we hit the right side of the range (nodes are sorted like a classical scanline algorithm).

I have a draft implementation of even simpler approach - small grid within tile that covers neighbors with cell size approx 300m x 300m, where all length of all edges of all nodes in each cell are summarazed. Then second pass creates second grid where each cell has sum of values from cells in radius. And then single lookup from node to the cell resolves the density value. So it requires just a single additional pass over all nodes to create such density index.
Just need to find some time to polish it and check how that grid approximation (all nodes within the same cell will have the same density) will change density distribution.